requirements.txt
Django==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==8.4.0
drf-yasg==1.21.7
django-cors-headers==4.2.0
mysqlclient==2.2.0